requirements.txt
awsebcli==3.7.3
beautifulsoup4==4.4.1
blessed==1.9.5
botocore==1.4.5
bs4==0.0.1
cement==2.4.0
colorama==0.3.3
Django==1.9.4
djangorestframework==3.3.3
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
jmespath==0.9.0
pathspec==0.3.3
python-dateutil==2.5.1
python-social-auth==0.2.14
PyYAML==3.11
requests==2.6.2
six==1.10.0
texttable==0.8.4
wcwidth==0.1.6
websocket-client==0.35.0
wheel==0.26.0
